+++ cogs/judge/__init__.py
@@ -1,6 +1,0 @@
-from .judge import Judge
-
-
-def setup(bot):
-    """Load cog"""
-    bot.add_cog(Judge(bot))