+++ cogs/review/__init__.py
@@ -1,0 +1,6 @@
+from .review import Review
+
+
+def setup(bot):
+    """Load cog"""
+    bot.add_cog(Review(bot))