+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Some cogs were renamed, you'll have to alter your main config.
 
 - Gatekeeper cog is renamed back to Verify
 - Faceshifter cog is renamed to Roles
+- Judge cog is renamed back to Review
 
 ## [1.1.0]